+++ tests/testthat/test-info_table.R
@@ -1,10 +1,17 @@
 test_that("exists", {
   expect_true(is.function(papercheck::info_table))
   expect_no_error(helplist <- help(info_table, papercheck))
+
+  expect_true(is.function(papercheck::author_table))
+  expect_no_error(helplist <- help(author_table, papercheck))
 })
 
 test_that("errors", {
   expect_error(info_table(),
+               "argument \"paper\" is missing, with no default",
+               fixed = TRUE)
+
+  expect_error(author_table(),
                "argument \"paper\" is missing, with no default",
                fixed = TRUE)
 })
@@ -36,4 +43,16 @@ test_that("missing items", {
   expect_equal(names(infotable), expected)
 
   expect_equal(infotable$`not a column`, rep(NA, length(paper)))
+})
+
+test_that("author_table", {
+  paper <- psychsci[[1]]
+  a1 <- author_table(paper)
+  expect_equal(nrow(a1), 7)
+  expect_equal(a1$n, 1:7)
+
+  paper <- psychsci[1:2]
+  a2 <- author_table(paper)
+  expect_equal(nrow(a2), 10)
+  expect_equal(a2$n, c(1:7, 1:3))
 })